+++ goodweproxy/goodwesensor.py
